+++ Stitch/Graph/Node/View/NodeView.swift
@@ -9,239 +9,6 @@ import SwiftUI
 import UIKit
 import StitchSchemaKit
 
-struct CacheData {
-    let maxSize: CGSize
-    let spacing: [CGFloat]
-    let totalSpacing: CGFloat
-}
-
-struct InfiniteCanvas: Layout {
-    let graph: GraphState
-    
-    typealias Cache = [CanvasItemId: CGRect]
-    
-    func sizeThatFits(proposal: ProposedViewSize, subviews: Subviews, cache: inout Cache) -> CGSize {
-        .init(width: proposal.width ?? .zero,
-              height: proposal.height ?? .zero)
-    }
-    
-    func placeSubviews(in bounds: CGRect, proposal: ProposedViewSize, subviews: Subviews, cache: inout Cache) {
-        // Place subviews
-        for subview in subviews {
-            // TODO: need support for removal
-            let positionData = subview[CanvasPositionKey.self]
-            let id = positionData.id
-            
-            guard let subviewSize = cache.get(id)?.size else {
-                // This is when with empty view for empty nodes
-                continue
-            }
-
-            subview.place(
-                at: bounds.origin,
-                anchor: .topLeading,
-                proposal: ProposedViewSize(subviewSize))
-        }
-    }
-    
-    func makeCache(subviews: Subviews) -> Cache {
-        // Only make cache when specified
-        if let existingCache = graph.visibleNodesViewModel.infiniteCanvasCache {
-            return existingCache
-        }
-        
-        let cache = subviews.reduce(into: Cache()) { result, subview in
-            let positionData = subview[CanvasPositionKey.self]
-            let id = positionData.id
-            let position = positionData.position
-            let size = subview.sizeThatFits(.unspecified)
-            
-            let bounds = CGRect(origin: position,
-                                size: size)
-            result.updateValue(bounds, forKey: id)
-        }
-        
-        graph.visibleNodesViewModel.infiniteCanvasCache = cache
-        return cache
-    }
-    
-    func updateCache(_ cache: inout Cache, subviews: Subviews) {
-        // Recalcualte graph if cache reset
-        if graph.visibleNodesViewModel.infiniteCanvasCache == nil {
-            cache = self.makeCache(subviews: subviews)
-            graph.visibleNodesViewModel.infiniteCanvasCache = cache
-        }
-    }
-    
-    func explicitAlignment(of guide: HorizontalAlignment,
-                           in bounds: CGRect,
-                           proposal: ProposedViewSize,
-                           subviews: Self.Subviews,
-                           cache: inout Cache) -> CGFloat? {
-        return nil
-    }
-    
-    func explicitAlignment(of guide: VerticalAlignment,
-                           in bounds: CGRect,
-                           proposal: ProposedViewSize,
-                           subviews: Self.Subviews,
-                           cache: inout Cache) -> CGFloat? {
-        return nil
-    }
-}
-
-struct CanvasLayoutPosition {
-    let id: CanvasItemId
-    let position: CGPoint
-}
-
-private struct CanvasPositionKey: LayoutValueKey {
-    static let defaultValue: CanvasLayoutPosition = .init(id: .node(.init()),
-                                                          position: .zero)
-}
-
-extension View {
-    func canvasPosition(id: CanvasItemId,
-                        position: CGPoint) -> some View {
-        layoutValue(key: CanvasPositionKey.self, value: .init(id: id,
-                                                              position: position))
-            .position(position)
-    }
-}
-
-protocol StitchLayoutCachable: AnyObject {
-    var viewCache: NodeLayoutCache? { get set }
-}
-
-struct NodeLayoutCache {
-    var sizes: [CGSize] = []
-    var sizeThatFits: CGSize = .zero
-    var spacing: ViewSpacing = .zero
-}
-
-struct NodeLayout<T: StitchLayoutCachable>: Layout {
-    typealias Cache = () -> Void //NodeLayoutCache
-    
-    let observer: T
-    
-    func sizeThatFits(proposal: ProposedViewSize, subviews: Subviews, cache: inout Cache) -> CGSize {
-        guard let cache = observer.viewCache else {
-            let newCache = self.createCache(subviews: subviews)
-            self.observer.viewCache = newCache
-            
-            return newCache.sizeThatFits
-        }
-        
-        return cache.sizeThatFits
-    }
-    
-    private func createCache(subviews: Subviews) -> NodeLayoutCache {
-        let sizes = subviews.map { $0.sizeThatFits(.unspecified) }
-        let sizeThatFits = self.calculateSizeThatFits(subviews: subviews)
-        let spacing = self.calculateSpacing(subviews: subviews)
-        let cache = NodeLayoutCache(sizes: sizes,
-                                    sizeThatFits: sizeThatFits,
-                                    spacing: spacing)
-        
-        return cache
-    }
-    
-    private func calculateSizeThatFits(subviews: Subviews) -> CGSize {
-        var totalWidth: CGFloat = 0
-        var totalHeight: CGFloat = 0
-        
-        for subview in subviews {
-            let subviewSize = subview.sizeThatFits(.unspecified) // Ask subview for its natural size
-            totalWidth = max(totalWidth, subviewSize.width) // Expand the width to fit the widest subview
-            totalHeight += subviewSize.height // Stack the subviews vertically
-        }
-        
-        return CGSize(width: totalWidth, height: totalHeight)
-    }
-    
-    func placeSubviews(in bounds: CGRect, proposal: ProposedViewSize, subviews: Subviews, cache: inout Cache) {
-        guard !subviews.isEmpty else { return }
-        
-        let cache = observer.viewCache ?? self.createCache(subviews: subviews)
-            
-        if self.observer.viewCache == nil {
-            self.observer.viewCache = cache
-        }
-        
-        for index in subviews.indices {
-            let subview = subviews[index]
-            
-            let size = cache.sizes[index]
-            
-//            let actualSize = subview.sizeThatFits(proposal)
-            
-            subview.place(
-                at: bounds.origin,
-                anchor: .topLeading,
-                proposal: ProposedViewSize(size))
-        }
-    }
-    
-    func spacing(subviews: Self.Subviews, cache: inout Cache) -> ViewSpacing {
-        guard let cache = self.observer.viewCache else {
-            let newCache = self.createCache(subviews: subviews)
-            return newCache.spacing
-        }
-        
-        return cache.spacing
-    }
-    
-    private func calculateSpacing(subviews: Self.Subviews) -> ViewSpacing {
-        var spacing = ViewSpacing()
-
-        for index in subviews.indices {
-            var edges: Edge.Set = [.leading, .trailing]
-            if index == 0 { edges.formUnion(.top) }
-            if index == subviews.count - 1 { edges.formUnion(.bottom) }
-            spacing.formUnion(subviews[index].spacing, edges: edges)
-        }
-
-        return spacing
-    }
-    
-    func makeCache(subviews: Subviews) -> Cache {
-//        guard let cache = observer.viewCache else {
-//            let sizes = subviews.map { $0.sizeThatFits(.unspecified) }
-//            let sizeThatFits = self.calculateSizeThatFits(subviews: subviews)
-//            let spacing = self.calculateSpacing(subviews: subviews)
-//            let cache = Cache(sizes: sizes,
-//                              sizeThatFits: sizeThatFits,
-//                              spacing: spacing)
-//            
-//            self.observer.viewCache = cache
-//            return cache
-//        }
-//        
-//        return cache
-        { }
-    }
-    
-    // Keep this empty for perf
-    func updateCache(_ cache: inout Cache, subviews: Subviews) { }
-    
-    func explicitAlignment(of guide: HorizontalAlignment,
-                           in bounds: CGRect,
-                           proposal: ProposedViewSize,
-                           subviews: Self.Subviews,
-                           cache: inout Cache) -> CGFloat? {
-        return nil
-    }
-    
-    func explicitAlignment(of guide: VerticalAlignment,
-                           in bounds: CGRect,
-                           proposal: ProposedViewSize,
-                           subviews: Self.Subviews,
-                           cache: inout Cache) -> CGFloat? {
-        return nil
-    }
-}
-
-
 struct NodeView<InputsViews: View, OutputsViews: View>: View {
     @Bindable var node: CanvasItemViewModel
     @Bindable var stitch: NodeViewModel
